% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmGetOrders.R
\name{mmGetOrders}
\alias{mmGetOrders}
\title{Getting Orders and 
ProjectId can be Vector!}
\usage{
mmGetOrders(acc_id = NULL, ProjectId = NULL, mm_token = new_token)
}
\description{
Getting Orders and 
ProjectId can be Vector!
}
