% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmGetPlacementInfo.R
\name{mmGetPlacementInfo}
\alias{mmGetPlacementInfo}
\title{Return placement info}
\usage{
mmGetPlacementInfo(acc_id = NULL, ProjectId = NULL,
  mm_token = new_token)
}
\arguments{
\item{acc_id}{}

\item{mm_token}{}
}
\description{
Return placement info
}
