% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmGetReportFields.R
\name{mmGetReportFields}
\alias{mmGetReportFields}
\title{mmGetReportFields}
\usage{
mmGetReportFields(acc_id = NULL, mm_token = new_token,
  ProjectId = NULL)
}
\arguments{
\item{ProjectId}{}
}
\description{
mmGetReportFields
}
