% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmGetStat.R
\name{mmGetStat}
\alias{mmGetStat}
\title{Function that get stats}
\usage{
mmGetStat(acc_id = NULL, ProjectId = NULL, mm_token = new_token,
  start_date = NULL, end_date = NULL, date_group = "month",
  group = "no")
}
\description{
Function that get stats
}
