% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmReportList.R
\name{mmReportList}
\alias{mmReportList}
\title{mmReportList
Getin last 100 reports}
\usage{
mmReportList(acc_id = NULL, ProjectId = NULL, mm_token = new_token)
}
\arguments{
\item{mm_token}{}
}
\description{
mmReportList
Getin last 100 reports
}
