% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaryR.R
\docType{package}
\name{mymaryR}
\alias{mymaryR}
\alias{mymaryR-package}
\title{honey: package to deal with good and bad bees}
\description{
Extremely important package for every Bear
}
\details{
I've written this package a long time ago in a far-far galaxy.
}
