% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmCrateReport.R
\name{mmCreateReport}
\alias{mmCreateReport}
\title{mmReportCreator}
\usage{
mmCreateReport(acc_id = NULL, ProjectId = NULL, mm_token = new_token,
  start_date = Sys.Date(), end_date = Sys.Date(), placements = NULL,
  dimensions = c("date", "ad_text", "keyword_text"),
  metrics = c("impressions", "clicks", "ctr", "cpc_fact", "cost_fact"),
  locale = "en", format = "XLSX")
}
\arguments{
\item{format}{}
}
\description{
mmReportCreator
}
