% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmReportList.R
\name{mmReportList}
\alias{mmReportList}
\title{mmReportList
Getin last 10 reports}
\usage{
mmReportList(acc_id = NULL, ProjectId = NULL, mm_token = new_token)
}
\arguments{
\item{mm_token}{}
}
\description{
Getin last 10 reports
}
\examples{
mmReportList(acc_id = 12345, ProjectId = 12345, mm_token = new_token)
}
